+++ integration/alertmanager_test.go
@@ -269,7 +269,7 @@ func TestAlertmanagerSharding(t *testing.T) {
 		require.NoError(t, am.WaitSumMetrics(e2e.Equals(float64(30)), "cortex_alertmanager_tenants_discovered"))
 		// We know that the ring has settled when every instance has some tenants and the total number of tokens have been assigned.
 		require.NoError(t, am.WaitSumMetrics(e2e.Greater(float64(0)), "cortex_alertmanager_tenants_owned"))
-		require.NoError(t, am.WaitSumMetrics(e2e.Equals(float64(384)), `cortex_ring_tokens_total`))
+		require.NoError(t, am.WaitSumMetrics(e2e.Equals(float64(384)), "cortex_ring_tokens_total"))
 	}
 
 	var totalTenants int